Move ChatWebSocketWorker logic into model modules
